Code-style

+++ Controllers/FSEGTObjectsContext.cpp
@@ -81,7 +81,7 @@ void FSEGTObjectsContext::subscribe(shared_ptr<FSEGTObjectContextDelegate> deleg
 
 void FSEGTObjectsContext::unsubscribe(shared_ptr<FSEGTObjectContextDelegate> delegate) {
     
-    for (int i = 0; i < subscribers.size(); i++) {
+    for (size_t i = 0; i < subscribers.size(); i++) {
         
         auto subscriber = subscribers.at(i);
